Litte refactor
Move the background mode logical to another class for better
code distribution.
